requirements.txt
jinja2
picamera
RPi.GPIO
